refactor(lib): move wallet api, and xorurl to their own files, plus remove unnecessary use clauses from several files
