+++ src/api/scl_mock.rs
@@ -6,9 +6,8 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::api::helpers::{
-    parse_hex, vec_to_hex, xorname_from_pk, xorname_to_hex, xorurl_to_xorname,
-};
+use super::xorurl::xorurl_to_xorname;
+use crate::api::helpers::{parse_hex, vec_to_hex, xorname_from_pk, xorname_to_hex};
 use log::debug;
 use rand::rngs::OsRng;
 use rand_core::RngCore;
